Dev #186: conversion of dataset and refactoring to allow preprocess-toolbox derived dataset training
